Added new infinite level and mechanic
Created a fourth level with an enemy spawner with basic configurable settings. The fourth level goes on infinitely. At the moment, if the EnemySpawner prefab is on any level, it will turn it into an infinite level. Some of the existing scripts are slightly modified to make the EnemySpawner work.